Add a header/footer to the Login modal popup
The content of the poopup is served from a different Zendesk domain to the rest of the help centre, so the styling of the popup itself cannot be altered.

It is possible to render an extra header and footer around the popup to provide supplementary content.

Currently it has been decided this provides no benefit.