Make #is conditions of sections use name rather than ID
Zendesk cannot compare string/number for equality - `{{#is 0 '0'}}` will never succeed.  The template variables must be strings, and the section IDs are numbers.  Therefore we cannot use the section IDs against variables configured per environment, so we must use the section names.

On the plus side, these do not need to be configured per environment as the sections have the same names in each.  On the other hand, it does restrict the section names to being unique.